Move application module out of TeamBuilder::main() function and into TeamBuilderApp
